speech-tools 1.2.96 (beta)
Directory: http://festvox.org/packed/festival/1.96/
File:      speech_tools-1.2.96-beta.tar.gz

+++ grammar/ngram/ngrammar_aux.cc
@@ -37,8 +37,8 @@
 /* Auxiliary functions for EST_Ngram class                               */
 /*                                                                       */
 /*=======================================================================*/
-#include <iostream.h>
-#include <string.h>
+#include <iostream>
+#include <cstring>
 #include "EST_String.h"
 #include "EST_Ngrammar.h"
 
@@ -110,7 +110,7 @@ smooth_ExponentialFit(EST_DVector &N, int first, int last)
     }
   
   for(int r=first;r<=last;r++)
-    N[r] = exp(a)* pow((float)r, b);
+    N[r] = exp(a)* pow((double)r, b);
   
   return true;
 }
@@ -252,7 +252,7 @@ void frequency_of_frequencies(EST_DVector &ff, EST_Ngrammar &n,int this_order)
 		  for (i=1;i<ff.n();i++)
 		    total += ff(i);
 		  
-		  ff[0] = pow(n.get_vocab_length(),n.order()) - total;
+		  ff[0] = pow(float(n.get_vocab_length()),float(n.order())) - total;
 	      }
 	}
 	break;
@@ -285,7 +285,7 @@ void frequency_of_frequencies(EST_DVector &ff, EST_Ngrammar &n,int this_order)
 		  double total=0;
 		  for (i=1;i<ff.n();i++)
 		    total += ff(i);
-		  ff[0] = pow(n.get_vocab_length(),this_order) - total;
+		  ff[0] = pow(float(n.get_vocab_length()),float(this_order)) - total;
